requirements.txt
asn1crypto==0.24.0
astroid==1.6.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.3.3
boto3==1.5.7
botocore==1.8.21
cffi==1.11.2
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
docutils==0.14
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
jieba==0.39
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
mysqlclient==1.3.12
parsel==1.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
PyDispatcher==2.0.5
pylint==1.8.1
pyOpenSSL==17.5.0
python-dateutil==2.6.1
queuelib==1.4.2
redis==2.10.6
s3transfer==0.1.12
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.2
Twisted==17.9.0
w3lib==1.18.0
wrapt==1.10.11
zope.interface==4.4.3
